Moving DiskBufferingConfiguration to the public api
