+++ agent-sdk/src/main/java/co/elastic/otel/android/features/diskbuffering/DiskBufferingConfiguration.kt
@@ -16,13 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.otel.android.internal.features.diskbuffering
+package co.elastic.otel.android.features.diskbuffering
 
 /**
- * This class is internal and is hence not for public use. Its APIs are unstable and can change at
- * any time.
+ * Defines the disk-buffering behavior.
  */
-data class DiskBufferingConfiguration private constructor(val enabled: Boolean) {
+class DiskBufferingConfiguration private constructor(val enabled: Boolean) {
     internal var maxCacheFileSize = 1024 * 1024
     internal var maxCacheSize = 15 * 1024 * 1024 // 15 MB
     internal var maxFileAgeForWrite: Long? = null
