+++ agent-sdk/src/test/java/co/elastic/otel/android/internal/features/persistence/DiskManagerTest.kt
@@ -18,7 +18,7 @@
  */
 package co.elastic.otel.android.internal.features.persistence
 
-import co.elastic.otel.android.internal.features.diskbuffering.DiskBufferingConfiguration
+import co.elastic.otel.android.features.diskbuffering.DiskBufferingConfiguration
 import co.elastic.otel.android.internal.features.diskbuffering.tools.DiskManager
 import co.elastic.otel.android.internal.services.appinfo.AppInfoService
 import io.mockk.every
